Added ability to type '/gms' for suvival and other gamemodes. Also tidied up some code.
