--- README.md.txt
@@ -1,11 +1,0 @@
-![foundation logo](https://www.vindere.us/img/header.png)
-
-![development notice](https://www.vindere.us/img/dev_notice.png)
-
-![description](https://www.vindere.us/img/description.png)
-
-![commands](https://www.vindere.us/img/command.png)
-
-![features](https://www.vindere.us/img/features.png)
-
-[![donate](https://www.vindere.us/img/donation.png)](http://www.paypal.me/SpriteV1dev)
