+++ src/main/java/us/vindere/foundation/commands/core/PingCommand.java
@@ -1,4 +1,4 @@
-package us.vindere.foundation.commands;
+package us.vindere.foundation.commands.core;
 
 public class PingCommand {
     //TODO ping command
