+++ src/main/java/us/vindere/foundation/commands/game/FeedCommand.java
@@ -1,4 +1,4 @@
-package us.vindere.foundation.commands;
+package us.vindere.foundation.commands.game;
 
 import org.bukkit.Bukkit;
 import org.bukkit.command.Command;
@@ -8,42 +8,44 @@ import org.bukkit.entity.Player;
 import org.bukkit.permissions.Permission;
 import us.vindere.foundation.Foundation;
 import us.vindere.foundation.placeholders.phError;
-import us.vindere.foundation.placeholders.commands.phHeal;
+import us.vindere.foundation.placeholders.commands.game.phFeed;
 import us.vindere.foundation.utils.Config;
 import us.vindere.foundation.utils.Tools;
 
-public class HealCommand implements CommandExecutor{
+public class FeedCommand implements CommandExecutor{
     Foundation main;
-    public HealCommand(Foundation instance) { main = instance; }
+    public FeedCommand(Foundation instance) { main = instance; }
     phError error = new phError();
-    phHeal heal = new phHeal();
+    phFeed feed = new phFeed();
 
     public boolean onCommand (CommandSender sender, Command cmd, String label, String[] args) {
         Config config = new Config(this);
         Player player = Tools.getPlayer(sender);
-        Permission selfHeal = new Permission("f.heal.self");
-        Permission othersHeal = new Permission("f.heal.others");
+        Permission selfFeed = new Permission("f.feed.self");
+        Permission othersFeed = new Permission("f.feed.others");
 
         if (player != null) {
             if (args.length < 1) {
-                if (sender.hasPermission(selfHeal)) {
-                    player.setHealth(20);
-                    heal.healed(main, sender);
+                if (sender.hasPermission(selfFeed)) {
+                    player.setExhaustion(0);
+                    player.setFoodLevel(20);
+                    feed.feeded(main, sender);
                     return true;
                 } else {
-                    error.noPerm(main, sender, selfHeal);
+                    error.noPerm(main, sender, selfFeed);
                     return false;
-                }
+            }
             } else if (args.length >= 1) {
                 Player player2 = Bukkit.getPlayer(args[0].toString());
                 if (player2 != null) {
-                    if (sender.hasPermission(othersHeal)) {
-                        player2.setHealth(20);
-                        heal.healer(main, sender, player2);
-                        heal.healee(main, sender, player2);
+                    if (sender.hasPermission(othersFeed)) {
+                        player2.setExhaustion(0);
+                        player2.setFoodLevel(20);
+                        feed.feeder(main, sender, player2);
+                        feed.feedee(main, sender, player2);
                         return true;
                     } else {
-                        error.noPerm(main, sender, othersHeal);
+                        error.noPerm(main, sender, othersFeed);
                         return false;
                     }
                 } else {
@@ -53,14 +55,15 @@ public class HealCommand implements CommandExecutor{
             }
         } else {
             if (args.length < 1) {
-                heal.consoleError(main, sender);
+                feed.consoleError(main, sender);
                 return false;
             } else if (args.length >= 1) {
                 Player player2 = Bukkit.getPlayer(args[0].toString());
                 if (player2 != null) {
-                    player2.setHealth(20);
-                    heal.healer(main, sender, player2);
-                    heal.healee(main, sender, player2);
+                    player2.setExhaustion(0);
+                    player2.setFoodLevel(20);
+                    feed.feeder(main, sender, player2);
+                    feed.feedee(main, sender, player2);
                     return true;
                 } else {
                     error.playerOffline(main, sender, args[0].toString());
