+++ src/main/java/us/vindere/foundation/placeholders/commands/admin/phVanish.java
@@ -1,0 +1,4 @@
+package us.vindere.foundation.placeholders.commands.admin;
+
+public class phVanish {
+}
