+++ src/main/java/us/vindere/foundation/placeholders/commands/core/phIdentity.java
@@ -1,4 +1,4 @@
-package us.vindere.foundation.placeholders.commands;
+package us.vindere.foundation.placeholders.commands.core;
 
 import org.bukkit.Bukkit;
 import org.bukkit.command.CommandSender;
